pyproject.toml
[tool.poetry]
name = "spht"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "3.7.4"
astroid = "2.4.2"
async-timeout = "3.0.1"
attrs = "20.3.0"
certifi = "2020.12.5"
chardet = "3.0.4"
colorama = "0.4.4"
"discord.py" = "1.6.0"
dotmap = "1.3.23"
idna = "2.10"
isort = "5.6.4"
lazy-object-proxy = "1.4.3"
mccabe = "0.6.1"
multidict = "4.7.6"
pandas = "1.1.5"
python-jsonstore = "1.3.0"
requests = "2.25.1"
requests-html = "0.10.0"
six = "1.15.0"
tabulate = "0.8.7"
toml = "0.10.2"
urllib3 = "1.26.3"
wrapt = "1.12.1"
yahoo-fin = "0.8.7"
yarl = "1.5.1"
wolframalpha = "^5.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"
pylint = "2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
